restore gameboy functionality (partial)
the rom still crashs once the entire titlescreen loads, but we are making progress. The issue was that Adam set it to always initialize fast CPU no matter what. I pulled a function from the LADX disassem, and managed to make it so that on a dmg, it loads the slower cpu speed. However, despite now loading, we insta crash as soon as the title graphics load.